Moved functionality from cdf/pdf to continuousDistribution
